agent/requirements.txt

##Agent Requirements
fastapi
uvicorn
mcp
google-generativeai
python-dotenv
